Add Set2a.hs
Solve Ex3
"The function takes an index i, a new value x, and the list xs to be updated as inputs. It uses the take function to get the elements up to the index i, then adds the new value x, and finally, uses the drop function to append the elements after index i. By doing this, it creates a new list with the element at the given index updated."